Ray collider code now is configurable: When you define OPC_RAYCOLLIDER_SCALE_BEFORE_OVERLAP, it tells opcode that the ray/line used by RayCollider is NOT in the model's local space. This means that model's AABBs/tris must be scaled before any intersection test takes place, which can be much slower for complex trees hit by the ray.
When this macro is not defined, the ray is transformed into the model's local space, so we don't 
need scale AABBs/tris before intersection tests. Because this, it could be very-very faster for
complex trees hit by the ray.

This macro allow us to compare performance between these techniques. Other thing it allow us to 
overcome a current bug when scaled meshes are used. Sooner or later, this code will be updated
and both options will become feasible and stable.

That's it!